Redesigned package structure
