+++ kryptonite/src/main/java/com/github/hpgrahsl/kryptonite/Kryptonite.java
@@ -16,6 +16,9 @@
 
 package com.github.hpgrahsl.kryptonite;
 
+import com.github.hpgrahsl.kryptonite.crypto.AesGcmNoPadding;
+import com.github.hpgrahsl.kryptonite.crypto.CryptoAlgorithm;
+import com.github.hpgrahsl.kryptonite.key.KeyVault;
 import java.util.Base64;
 import java.util.LinkedHashMap;
 import java.util.Map;
