+++ kryptonite/src/main/java/com/github/hpgrahsl/kryptonite/key/KeyException.java
@@ -14,26 +14,26 @@
  * limitations under the License.
  */
 
-package com.github.hpgrahsl.kryptonite;
+package com.github.hpgrahsl.kryptonite.key;
 
 @SuppressWarnings("serial")
-public class KeyInvalidException extends KeyException {
+public class KeyException extends RuntimeException {
 
-  public KeyInvalidException() {}
+  public KeyException() {}
 
-  public KeyInvalidException(String message) {
+  public KeyException(String message) {
     super(message);
   }
 
-  public KeyInvalidException(String message, Throwable cause) {
+  public KeyException(String message, Throwable cause) {
     super(message, cause);
   }
 
-  public KeyInvalidException(Throwable cause) {
+  public KeyException(Throwable cause) {
     super(cause);
   }
 
-  public KeyInvalidException(
+  public KeyException(
       String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
     super(message, cause, enableSuppression, writableStackTrace);
   }
